+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.10.1] - 2026-04-14
+
+### Added
+- `XMLNode.remove_child()` method for removing direct child elements from the XML tree.
+
+### Testing
+- Added unit tests for `XMLNode.remove_child()`.
+
+
 ## [0.10.0] - 2026-04-11
 
 ### Added